Moved common test code into test/common.rs.  Changed FakeFdOps to MockFdOps.
